Add an action and a reducer for saving the form
Why:

* When the Quote Rates button is clicked, the API should be called with
the data input by the user. Only validated data should be sent to the
API. To ensure this, the unvalidated form data will be stored in local
component state. When the Quote Rates button is clicked, if the data is
valid, it will be saved to Redux state. The API will then be called with
the data in Redux state. In order to save the data to Redux state, an
action and a reducer are required.

This change addresses the need by:

* Create a constant for the "SAVE_FORM" action type
* Create and test an action with a payload of form data
* Create and test a reducer that updates Redux state with form data